Added properties to fetch repos, readme and contributions
