+++ src/scrape_up/github/repo.py
@@ -1,0 +1,7 @@
+import requests
+from bs4 import BeautifulSoup
+
+class Repository:
+    """A python wrapper to scrape any github repository"""
+    def __init__(self, url):
+        pass
